+++ lib/chef/provider/package/portage.rb
@@ -110,9 +110,7 @@ class Chef
             pkg = "~#{name}-#{$1}"
           end
 
-          run_command_with_systems_locale(
-            :command => "emerge -g --color n --nospinner --quiet#{expand_options(@new_resource.options)} #{pkg}"
-          )
+          shell_out!( "emerge -g --color n --nospinner --quiet#{expand_options(@new_resource.options)} #{pkg}" )
         end
 
         def upgrade_package(name, version)
@@ -126,9 +124,7 @@ class Chef
             pkg = "#{@new_resource.package_name}"
           end
 
-          run_command_with_systems_locale(
-            :command => "emerge --unmerge --color n --nospinner --quiet#{expand_options(@new_resource.options)} #{pkg}"
-          )
+          shell_out!( "emerge --unmerge --color n --nospinner --quiet#{expand_options(@new_resource.options)} #{pkg}" )
         end
 
         def purge_package(name, version)