+++ lib/chef/provider/service/simple.rb
@@ -85,16 +85,16 @@ class Chef
         end
 
         def start_service
-          shell_out!(@new_resource.start_command)
+          shell_out_with_systems_locale!(@new_resource.start_command)
         end
 
         def stop_service
-          shell_out!(@new_resource.stop_command)
+          shell_out_with_systems_locale!(@new_resource.stop_command)
         end
 
         def restart_service
           if @new_resource.restart_command
-            shell_out!(@new_resource.restart_command)
+            shell_out_with_systems_locale!(@new_resource.restart_command)
           else
             stop_service
             sleep 1
@@ -103,7 +103,7 @@ class Chef
         end
 
         def reload_service
-          shell_out!(@new_resource.reload_command)
+          shell_out_with_systems_locale!(@new_resource.reload_command)
         end
 
       protected